// src/api/generated/types.gen.ts
export type UserAddressDto = {
    /**
     * Unique address CUID
     */
    id: string;
    /**
     * Address type
     */
    type: 'SHIPPING' | 'BILLING';
    /**
     * Whether this is the default address for its type
     */
    isDefault: boolean;
    /**
     * Recipient full name
     */
    fullName: string;
    /**
     * Contact phone number
     */
    phone: string;
    /**
     * Street address
     */
    street: string;
    /**
     * City name
     */
    city: string;
    /**
     * Region or state
     */
    region?: {
        [key: string]: unknown;
    };
    /**
     * Postal code
     */
    postalCode: string;
    /**
     * ISO 3166-1 alpha-2 country code
     */
    country: string;
    createdAt: string;
    updatedAt: string;
};